requirements.txt
fastapi
uvicorn
python-dotenv
langchain 
langsmith 
langchain_benchmarks 
openai 
chromadb 
pypdfium2 
open-clip-torch 
pillow
langchain_experimental
langchain_chroma
langchain_openai
langchain_community 
sentence-transformers
rerankers
imgbeddings
pypdf
sentence_transformers
pymupdf
pinecone
streamlit
groq
sounddevice
soundfile
spacy
pytesseract
bs4
markdown
matplotlib
streamlit_mermaid
llama-index
llama_index.legacy
cohere
llama-index-node-parser-topic
langgraph
langsmith